requirements.txt
django~=4.2
pillow
django-crispy-forms
crispy-bootstrap5
jdatetime
requests
sorl-thumbnail
django-qsessions
django-xforwardedfor-middleware
django-axes[ipware]
django_ratelimit
django-auditlog
django-cachalot
#django-recaptcha
django-turnstile
django-ipware
user_agents
django-serverside-datatable
valkey[libvalkey]
django-valkey[libvalkey]
django-prometheus
gunicorn
psycopg